Move docker_autoclean to end
By running this first, it was common that all docker images for this
project would be removed and thus any commands which may have hit the
build cache would need to start from scratch

By moving the autoclean call to the end, we can make use of any cached
images and the proceed to cleanup any unused. This menas that any common
base layers will also be kept after the cleanup.